% Generated by roxygen2 (4.0.2): do not edit by hand
\name{grabtikz}
\alias{grabtikz}
\title{This is just a place holder right now}
\usage{
grabtikz(x = "")
}
\arguments{
\item{X}{character value, optional argument}
}
\description{
This pic does nothing right now now tikz yet
}
\examples{
grabtikz('send me your tikpics')
}

