% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tikpics}
\alias{tikpics}
\title{This is just a place holder right now}
\usage{
tikpics(x, y, z = NULL, data = NULL, geom = "scatter", xlab = NULL,
  ylab = NULL, zlab = NULL, tour = FALSE)
}
\description{
This pic does nothing right now now tikz yet
}
\examples{
tikpics(x,y,z)
}

