% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tikpics2Dscatter}
\alias{tikpics2Dscatter}
\title{This is just a place holder right now}
\usage{
tikpics2Dscatter(options = NULL)
}
\arguments{
\item{options}{character value, optional argument}
}
\description{
This pic does nothing right now now tikz yet
}
\examples{
tikpics2Dscatter('send me your tikpics')
}

