% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tikpics3Dscatter}
\alias{tikpics3Dscatter}
\title{This is just a place holder right now}
\usage{
tikpics3Dscatter(options = NULL)
}
\arguments{
\item{options}{character value, optional argument}
}
\description{
This pic does nothing right now now tikz yet
}
\examples{
tikpics3Dscatter('send me your tikpics')
}

